+++ app/scripts/Storage.js
