+++ app/scripts/actions/index.js
@@ -1,0 +1,11 @@
+import addOldArticleAlert from './add-old-article-alert'
+import addStockButtonToHeader from './add-stock-button-to-header'
+import autoEnableSyncScroll from './auto-enable-sync-scroll'
+import fixHeader from './fix-header'
+
+export default {
+  addOldArticleAlert,
+  addStockButtonToHeader,
+  autoEnableSyncScroll,
+  fixHeader
+}
