+++ app/scripts/contentsscript.js
@@ -1,7 +1,18 @@
-import isTeamPage from './modules/isTeamPage'
-import runActions from './modules/runActions'
+import isTeamPage from './isTeamPage'
+import actions from './actions'
+import fetchConfigOrDefault from './fetchConfigOrDefault'
 
 if (isTeamPage(document)) {
   console.log('Qiita Team Plusを実行します。')
-  runActions()
+  Object.values(actions)
+    .filter(({ routes }) => routes.some(route => window.location.pathname.match(route)))
+    .forEach(action => {
+      fetchConfigOrDefault(action.key)
+        .then(enabled => {
+          if (enabled) {
+            console.log(`アクション「${action.name}」を実行します。`)
+            action.run()
+          }
+        })
+    })
 }
