+++ app/scripts/defaultConfig.json
