+++ app/scripts/fetchConfigOrDefault.js
