+++ app/scripts/isTeamPage.js
