--- app/scripts/modules/getActions.js
@@ -1,9 +1,0 @@
-import routes from './routes'
-
-function getActions (path) {
-  return routes
-    .filter(route => path.match(route.path))
-    .flatMap(route => route.actions)
-}
-
-export default getActions
