+++ app/scripts/paths.js
